+++ NEWS.md
@@ -9,7 +9,7 @@
 
 ### Breaking Change
 
-* `qenv` objects should now be created with `qenv()` rather than `new_qenv()`. The new constructor always creates an empty object. `new_qenv` is now deprecated.
+* `qenv` objects should now be created with `qenv()` rather than `new_qenv()` (which has been removed). The new constructor always creates an empty object.
 
 ### Miscellaneous